pyproject.toml
[project]
name = "pipepost"
version = "0.1.0"
description = "Open-source AI content curation pipeline — scout, translate, publish"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [{name = "Denis Sultanov", email = "denis@denishub.dev"}]
keywords = [
    "ai",
    "content",
    "curation",
    "translation",
    "pipeline",
    "automation",
    "rss",
    "llm",
    "publishing",
    "news",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "litellm>=1.30",
    "defusedxml>=0.7",
    "duckduckgo-search>=6.0",
    "markdownify>=0.14",
    "tenacity>=8.2",
]

[project.optional-dependencies]
metrics = ["prometheus-client>=0.20"]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "respx",
    "pytest-cov",
    "types-PyYAML",
    "types-beautifulsoup4",
    "types-defusedxml",
    "prometheus-client>=0.20",
]

[project.urls]
Homepage = "https://github.com/densul/pipepost"
Repository = "https://github.com/densul/pipepost"
Issues = "https://github.com/densul/pipepost/issues"
Documentation = "https://github.com/densul/pipepost#readme"
Changelog = "https://github.com/densul/pipepost/releases"

[project.scripts]
pipepost = "pipepost.cli:main"

[tool.setuptools.packages.find]
include = ["pipepost*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
